Simple persistance in state ~ Important
This is a simple persistance to the state. I handle this by making a subscription to the store, I save the store when it changes using localStorage.
I'm not 100% sure if this is a good solution.
But right now it works and I can take the localStorage state to a persistent file using "fs".
I found that there's a thing called "redux-persist". Maybe I have to research that, but for now this works